[TR-207] Move route-linked data from timeline-configuration to route-data
